Slightly more well behaved

+++ hide.py
@@ -11,7 +11,7 @@ import json
 
 class UpdateErrorPanelCommand(sublime_plugin.TextCommand):
     def run(self, edit, errors):
-        self.view.insert(edit, 0, str(errors))
+        self.view.insert(edit, 0, str(errors) + "\n\n")
 
 class IdeBackendSaveListener(sublime_plugin.EventListener):
     def on_post_save(self, view):
@@ -31,14 +31,14 @@ def send_request(view, request):
 
 class IdeBackendAutocompleteHandler(sublime_plugin.EventListener):
     def on_query_completions(self, view, prefix, locations):
-      request = {
-        "request":"getAutocompletion", 
-        "autocomplete": {
-          "filePath": view.file_name(), 
-          "prefix": prefix
-          } 
-        }
-      send_request(view, request)
+        request = {
+            "request":"getAutocompletion", 
+            "autocomplete": {
+                "filePath": view.file_name(), 
+                "prefix": prefix
+                } 
+            }
+        send_request(view, request)
 
 class SendIdeBackendRequestCommand(sublime_plugin.WindowCommand):
     """
@@ -46,37 +46,60 @@ class SendIdeBackendRequestCommand(sublime_plugin.WindowCommand):
     (Sublime Text uses the class name to determine the name of the command
     the class executes when called, i.e. send_ide_backend_request)
     """
+
+    backendsByWindow = {}
+
     def __init__(self, window):
         super(SendIdeBackendRequestCommand, self).__init__(window)
         self.boot_ide_backend()
 
     def first_folder(self):
-      return self.window.folders()[0]
+        return self.window.folders()[0]
 
     def boot_ide_backend(self):
-      """
-      Start up a ide-backend-client subprocess for the window, and a thread to consume its stdout
-      """
-      print("Launching HIDE in " + self.first_folder())
+        """
+        Start up a ide-backend-client subprocess for the window, and a thread to consume its stdout
+        """
+        print("Launching HIDE in " + self.first_folder())
+        print("Backends by window: " + str(self.backendsByWindow))
+        self.backendsByWindow[self.window.id()] = "Hi!"
+ 
+ 
+        self.process = subprocess.Popen(["/Users/lukexi/.cabal/bin/ide-backend-client", "cabal", "."],
+            stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE,
+            cwd=self.first_folder()
+            )
+        
+        self.stdoutThread = threading.Thread(target=self.read_stdout)
+        self.stdoutThread.start()
+        
+        # self.stderrThread = threading.Thread(target=self.read_stderr)
+        # self.stderrThread.start()
 
-      self.process = subprocess.Popen(["/Users/lukexi/.cabal/bin/ide-backend-client", "cabal", "."],
-         stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.STDOUT,
-         cwd=self.first_folder()
-         )
-      
-      self.stdoutThread = threading.Thread(target=self.read_stdout)
-      self.stdoutThread.start()
+    def read_stderr(self):
+        while self.process.poll() is None:
+            try:
+                print("STDERR:")
+                print(self.process.stderr.readline().decode('UTF-8'))
+            except:
+                print("stderr process ending due to exception:", str(sys.exc_info()))
+                return;
+        print("process ended...")
 
     def read_stdout(self):
         while self.process.poll() is None:
             try:
-                data = json.loads(self.process.stdout.readline().decode('UTF-8'))
+                raw = self.process.stdout.readline().decode('UTF-8')
+
+                data = json.loads(raw)
                 
-                # print(json.loads(data))
+                print(data)
                 progress = data.get("progress")
                 response = data.get("response")
                 if progress:
-                  sublime.status_message(data["progress"]["parsedMsg"])
+                  progressMessage = progress.get(parsedMsg)
+                  if progressMessage:
+                    sublime.status_message(progressMessage)
                 elif response == "getSourceErrors":
                   errors = data.get("errors")
                   if errors:
@@ -85,40 +108,45 @@ class SendIdeBackendRequestCommand(sublime_plugin.WindowCommand):
                   print(data)
                 
             except:
-                print("process ending due to exception:", str(sys.exc_info()[0]))
-                # sublime.set_timeout(self.boot_ide_backend, 0)
+                print("process ending due to exception:", str(sys.exc_info()))
+                self.process.terminate()
+                self.process = None
                 return;
         print("process ended...")
 
-    def highlight_errors(self, errors):
-      error_panel = self.window.create_output_panel("hide_errors")
-      # error_panel.set_read_only(True)
-      # error_panel.set_scratch(True)
-      self.window.run_command("show_panel", {"panel":"output.hide_errors"})
 
-      for error in errors:
-        msg = error.get("msg")
-        error_panel.run_command("update_error_panel", {"errors":msg})
-        span = error.get("span")
-        kind = error.get("kind")
-        regionAndFileView = regionFromJSON(span, self.window)
-        print(str(regionAndFileView))
-        if regionAndFileView:
-          (region, file_view) = regionAndFileView
-          file_view.add_regions(str(region), [region], "invalid", "dot", sublime.DRAW_OUTLINED)
+    def highlight_errors(self, errors):
+        error_panel = self.window.create_output_panel("hide_errors")
+        # error_panel.set_read_only(True)
+        # error_panel.set_scratch(True)
+        self.window.run_command("show_panel", {"panel":"output.hide_errors"})
+
+        for error in errors:
+            msg = error.get("msg")
+            error_panel.run_command("update_error_panel", {"errors":msg})
+            span = error.get("span")
+            # kind = error.get("kind")
+            regionAndFileView = regionFromJSON(span, self.window)
+            print(str(regionAndFileView))
+            if regionAndFileView:
+                (region, file_view) = regionAndFileView
+                print("Adding error at "+ str(span) +": " + str(msg))
+                file_view.add_regions(str(region), [region], "invalid", "dot", sublime.DRAW_OUTLINED)
 
     def run(self, request):
         """
         Called via run_command("send_ide_backend_request", {"request":})
         """
-        # request = {"request":"getSourceErrors"}
-        print("SENDING " + str(request))
-        encodedString = json.JSONEncoder().encode(request) + "\n"
-        self.process.stdin.write(bytes(encodedString, 'UTF-8'))
-        self.process.stdin.flush()
+        if self.process:
+            print("SENDING " + str(request))
+            encodedString = json.JSONEncoder().encode(request) + "\n"
+            self.process.stdin.write(bytes(encodedString, 'UTF-8'))
+            self.process.stdin.flush()
 
     def __del__(self):
-        self.process.terminate()
+        if self.process:
+            self.process.terminate()
+            self.process = None
 
 def regionFromJSON(span, window):
   if span == None:
